Move Docker handling to dedicated module
