+++ emulator.cabal
@@ -105,6 +105,7 @@ test-suite emulator-tests
         Test.OnDemand
         Test.Warden
         Test.Utils.OnDemand
+        Test.Utils.Docker
     type:    exitcode-stdio-1.0
     build-depends:
         base
